% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb.Booster.R
\name{print.xgb.Booster}
\alias{print.xgb.Booster}
\title{Print xgb.Booster}
\usage{
\method{print}{xgb.Booster}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{an xgb.Booster object}

\item{verbose}{whether to print detailed data (e.g., attribute values)}

\item{...}{not currently used}
}
\description{
Print information about xgb.Booster.
}
\examples{
### Save code in a file "demo.r" and run with 1 processors by
### SHELL> mpiexec -np 1 Rscript demo.r

spmd.code <- "
### Initial.
suppressMessages(library(pbdMPI, quietly = TRUE))
suppressMessages(library(pbdXGB, quietly = TRUE))
init()

### Examples.
data(agaricus.train, package='xgboost')
train <- agaricus.train
bst <- xgboost(data = train$data, label = train$label, max_depth = 2,
               eta = 1, nthread = 2, nrounds = 2, objective = 'binary:logistic')
attr(bst, 'myattr') <- 'memo'

print(bst)
print(bst, verbose=TRUE)

### Finish.
finalize()
"
pbdMPI::execmpi(spmd.code = spmd.code, nranks = 1L)

}
