% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rabit.R
\name{xgb.rabit.init}
\alias{xgb.rabit.init}
\alias{xgb.rabit.finalize}
\alias{xgb.rabit.getworldsize}
\alias{xgb.rabit.getrank}
\title{Simple interface for rabit MPI C calls.}
\usage{
xgb.rabit.init()

xgb.rabit.finalize()

xgb.rabit.getworldsize()

xgb.rabit.getrank()
}
\description{
These are used to interface with MPI calls.
}
\details{
See MPI library for details or \pkg{pbdMPI} for examples.
}
